+++ test/runner.py
@@ -273,11 +273,10 @@ def choose_random_tests(base, num_tests, relevant_modes):
     chosen.add(new_test)
     if len(chosen) > before:
       print('* ' + new_test)
-    else:
-      # we may have hit the limit
-      if len(chosen) == len(tests) * len(relevant_modes):
-        print(f'(all possible tests chosen! {len(chosen)} = {len(tests)}*{len(relevant_modes)})')
-        break
+    elif len(chosen) == len(tests) * len(relevant_modes):
+      # we hit the limit
+      print(f'(all possible tests chosen! {len(chosen)} = {len(tests)}*{len(relevant_modes)})')
+      break
   return list(chosen)
 
 
@@ -421,10 +420,9 @@ def load_test_suite(args, modules, options):
       is_parallel_module = use_parallel_suite(m)
       if using_parallel_suite is None:
         using_parallel_suite = is_parallel_module
-      else:
+      elif is_parallel_module != using_parallel_suite:
         # All the following modules must match in their support for the parallel runner.
-        if is_parallel_module != using_parallel_suite:
-          utils.exit_with_error(f'attempt to mix parallel and non-parallel test modules ({m.__name__})')
+        utils.exit_with_error(f'attempt to mix parallel and non-parallel test modules ({m.__name__})')
 
   # If we are only running a single tests, never use the parallel tests suite.
   # This means that the output of a single test is always going to be in `out/test/` rather
@@ -542,7 +540,7 @@ def parse_args():
 
   if options.ansi is None:
     options.ansi = colored_logger.ansi_color_available()
-  else:
+  else:  # ruff: ignore[collapsible-else-if]
     if options.ansi:
       colored_logger.enable(force=True)
     else:
@@ -604,14 +602,14 @@ def log_test_environment():
   print(f'Python: "{sys.executable}". Version: {sys.version}')
   print(f'Emscripten test runner path: "{os.path.realpath(__file__)}"')
 
-  if os.path.isdir(utils.path_from_root('.git')):
+  if os.path.exists(utils.path_from_root('.git')):
     print(f'\nEmscripten repository: "{__rootpath__}"')
 
   emscripten_version = utils.path_from_root('emscripten-version.txt')
   if os.path.isfile(emscripten_version):
     print(f'emscripten-version.txt: {utils.EMSCRIPTEN_VERSION}')
 
-  if os.path.isdir(os.path.join(__rootpath__, '.git')):
+  if os.path.exists(os.path.join(__rootpath__, '.git')):
     print_repository_info(__rootpath__, 'Emscripten')
 
   print(f'EM_CONFIG: "{config.EM_CONFIG}"')
@@ -630,7 +628,7 @@ def log_test_environment():
   # Detect emsdk directory structure (build root vs source root)
   if re.match(r'main_.*_64bit_binaryen', os.path.basename(binaryen_git_dir)):
     binaryen_git_dir = os.path.realpath(os.path.join(binaryen_git_dir, '..', 'main'))
-  if os.path.isdir(os.path.join(binaryen_git_dir, '.git')):
+  if os.path.exists(os.path.join(binaryen_git_dir, '.git')):
     print(f'Binaryen git directory: "{binaryen_git_dir}"')
     print_repository_info(binaryen_git_dir, 'Binaryen')
 
@@ -639,10 +637,10 @@ def log_test_environment():
   # Find LLVM git directory in emsdk aware fashion
   def find_llvm_git_root(dir):
     while True:
-      if os.path.isdir(os.path.join(dir, ".git")):
+      if os.path.exists(os.path.join(dir, '.git')):
         return dir
-      if os.path.isdir(os.path.join(dir, "src", ".git")):
-        return os.path.join(dir, "src")
+      if os.path.exists(os.path.join(dir, 'src', '.git')):
+        return os.path.join(dir, 'src')
       if os.path.dirname(dir) == dir:
         return None
       dir = os.path.dirname(dir)
@@ -672,7 +670,7 @@ def log_test_environment():
   emsdk_dir = os.getenv('EMSDK')
   print(f'\nEMSDK: "{emsdk_dir}"')
   if emsdk_dir:
-    if os.path.isdir(os.path.join(emsdk_dir, '.git')):
+    if os.path.exists(os.path.join(emsdk_dir, '.git')):
       print_repository_info(emsdk_dir, 'Emsdk')
 
   print('==================== End of Test Setup =====================')
@@ -781,14 +779,14 @@ if __name__ == '__main__':
   except KeyboardInterrupt:
     logger.warning('KeyboardInterrupt')
     sys.exit(1)
-else:
-  # We are not the main process, and most likely a child process of
-  # the multiprocess pool.  In this mode the modifications made to the
-  # test class by `skip_test` need to be re-applied in each child
-  # subprocess (sad but true).  This is needed in particular on macOS
-  # and Windows where the default mode for multiprocessing is `spawn`
-  # rather than `fork`
-  if 'EMTEST_SKIP' in os.environ:
-    modules = get_and_import_modules()
-    for skip in os.environ['EMTEST_SKIP'].split():
-      skip_test(skip, modules)
+
+# We are not the main process, and most likely a child process of
+# the multiprocess pool.  In this mode the modifications made to the
+# test class by `skip_test` need to be re-applied in each child
+# subprocess (sad but true).  This is needed in particular on macOS
+# and Windows where the default mode for multiprocessing is `spawn`
+# rather than `fork`
+if 'EMTEST_SKIP' in os.environ:
+  modules = get_and_import_modules()
+  for skip in os.environ['EMTEST_SKIP'].split():
+    skip_test(skip, modules)